// apps/frontend/src/api/generated/model/publicationReportDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * QuizDock API
 * API REST du builder de quiz et des restitutions
 * OpenAPI spec version: 0.1.0
 */
import type { PublicationReportDtoHeaviestItem } from './publicationReportDtoHeaviestItem';
import type { PublicationReportDtoIssuesItem } from './publicationReportDtoIssuesItem';
import type { PublicationReportDtoUncreditedItem } from './publicationReportDtoUncreditedItem';

export interface PublicationReportDto {
  slug: string;
  slugSet: boolean;
  language: string;
  /** @nullable */
  license: string | null;
  tags: string[];
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  estimatedBytes: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  maxBytes: number;
  issues: PublicationReportDtoIssuesItem[];
  heaviest: PublicationReportDtoHeaviestItem[];
  uncredited: PublicationReportDtoUncreditedItem[];
}